requirements.txt
pytesseract
PyPDF2
Pillow
